chore: 瓦片代码重构
