+++ dev-demos/features/tile/vector.md
@@ -1,13 +1,13 @@
 ### Vector
-
-<!-- #### vector point
-<code src="./vector/point.tsx"></code>
+<!-- 
+#### vector point
+<code src="./vector/point.tsx"></code> -->
 
 #### vector line
-<code src="./vector/line.tsx"></code> -->
+<code src="./vector/line.tsx"></code>
 
-#### geojson-vt
-<code src="./vector/geojson-vt.tsx"></code>
+<!-- #### geojson-vt
+<code src="./vector/geojson-vt.tsx"></code> -->
 
 <!-- #### debugLayer
 <code src="./vector/debugLayer.tsx"></code> -->
