+++ packages/layers/src/tile/tileFactory/rasterData.ts
@@ -1,7 +1,6 @@
 import { ILayer, ISubLayerInitOptions } from '@antv/l7-core';
 import { Tile } from '@antv/l7-utils';
 import { ITileFactoryOptions } from '../interface';
-import { registerLayers } from '../utils';
 import TileFactory from './base';
 import RasterDataLayer from './layers/rasterDataLayer';
 
@@ -53,8 +52,6 @@ export default class RasterTiffTile extends TileFactory {
         clampLow,
       });
     this.emitEvent([layer], false);
-
-    registerLayers(this.parentLayer, [layer]);
 
     return {
       layers: [layer],
